+++ src/TRegx/CleanRegex/Match/Offset/MatchOffsetLimit.php
@@ -2,6 +2,7 @@
 namespace TRegx\CleanRegex\Match\Offset;
 
 use InvalidArgumentException;
+use Iterator;
 use TRegx\CleanRegex\Exception\GroupNotMatchedException;
 use TRegx\CleanRegex\Exception\NonexistentGroupException;
 use TRegx\CleanRegex\Exception\SubjectNotMatchedException;
@@ -12,7 +13,7 @@ use TRegx\CleanRegex\Internal\Match\Stream\Stream;
 use TRegx\CleanRegex\Match\FluentMatchPattern;
 use TRegx\CleanRegex\Match\Groups\Strategy\MatchAllGroupVerifier;
 
-class MatchOffsetLimit implements OffsetLimit, Stream
+class MatchOffsetLimit implements OffsetLimit, Stream, \IteratorAggregate
 {
     /** @var Base */
     private $base;
@@ -66,6 +67,11 @@ class MatchOffsetLimit implements OffsetLimit, Stream
         return $matches->getLimitedGroupOffsets($this->nameOrIndex, -1);
     }
 
+    public function getIterator(): Iterator
+    {
+        return new \ArrayIterator($this->all());
+    }
+
     /**
      * @param int $limit
      * @return (int|null)[]
@@ -87,7 +93,7 @@ class MatchOffsetLimit implements OffsetLimit, Stream
         return new FluentMatchPattern($this, new NotMatchedFluentOptionalWorker(new NoFirstElementFluentMessage()));
     }
 
-    public function firstKey()
+    public function firstKey(): int
     {
         return 0;
     }